requirements.txt
opencv-python==4.5.1.48
numpy
matplotlib
scipy
pillow
scikit-learn
torch
torchvision
torchaudio
tqdm
torchsummary 